[dart2wasm] Add --dry-run flag to dart2wasm.
This new flag will run the CFE to create a kernel and then run a series of checks over the resulting kernel to look for errors that could block a wasm migration.

The compiler will then exit before actually starting the wasm compilation process. This means no output file will be emitted so callers must be aware of this.

This first CL is not meant to cover every check we could add here. It adds some initial checks and we can expand on this to include more in follow-up changes.

One of the checks implemented here is also provided by a lint. While ideally we would share code between lints and these checks, the delta in the CFE vs analyzer model makes that infeasible today.

Sample output:
```
Found incompatibilities with WebAssembly.

package:dryrun/test.dart 5:15 - Cannot test a JS value against String (3)
package:dryrun/test.dart 6:7 - JS interop class 'B' cannot extend Dart class 'A'. (2)
```

Bug: https://github.com/dart-lang/sdk/issues/60050
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/437960
Commit-Queue: Nate Biggs <natebiggs@google.com>

+++ pkg/dartdev/lib/src/commands/compile.dart
@@ -1031,6 +1031,11 @@ class CompileWasmCommand extends CompileSubcommandCommand {
 
     final bool strip = args.flag('strip-wasm');
 
+    // When running in dry run mode there will not be any file emitted.
+    final isDryRun = extraCompilerOptions.any((e) => e.contains('dry-run'));
+
+    if (isDryRun) return 0;
+
     if (runWasmOpt) {
       final unoptFile = '$outputFileBasename.unopt.wasm';
       File(outputFile).renameSync(unoptFile);
